GITHUB: Added working UDP. Fixed README example.
Added test_udp.cpp as a udp example.
Added SendTo and RecvFrom due to being common functions used in udp socket examples.
Added Cache off to cmakelists.txt so testing can actually be turned off and on.
Fixed the example in the readme so it actually compiled.
Added install targets to main CMakeLists.txt so this could be used as a
  cmake external project
Added preprocessor system checking so IDEs can error check correctly.

Signed-off-by: Josh Miklos <josh.miklos@hobovrlabs.org>